demo: rename drm.js to parseDRMConfigurations.js
