+++ demo/full/scripts/lib/parseDRMConfigurations.js
